Update migration files to enhance database schema integrity by modifying storage field type in subscription plans, adjusting foreign key constraints for users and roles, and implementing unique indexes for better handling of nullable organization IDs. Additionally, update user roles in seeders for improved clarity and consistency.

+++ database/migrations/2025_08_22_143737_create_subscription_plans_table.php
@@ -30,7 +30,7 @@ return new class extends Migration
             $table->integer('max_knowledge_articles')->default(100);
             $table->integer('max_monthly_messages')->default(1000);
             $table->integer('max_monthly_ai_requests')->default(100);
-            $table->integer('max_storage_gb')->default(1);
+            $table->decimal('max_storage_gb', 5, 2)->default(1.00);
             $table->integer('max_api_calls_per_day')->default(1000);
 
             // Feature Flags
